add rendezvousHashring
