+++ pkg/receive/config.go
@@ -144,7 +144,7 @@ type ShuffleShardingConfig struct {
 	// across the available zones, but we don't try to balance the shards across zones.
 	ZoneAwarenessDisabled bool `json:"zone_awareness_disabled"`
 	// AlignedOrdinalSharding enables aligned shard selection for shuffle sharding.
-	// When true and using aligned_ketama algorithm, the same shards are selected
+	// When true and using rendezvous algorithm, the same shards are selected
 	// across all AZs, preserving strict replica alignment. ShardSize represents the
 	// number of shards to select (resulting in ShardSize * numAZs total endpoints).
 	AlignedOrdinalSharding bool                            `json:"aligned_ordinal_sharding"`
